[#0] Promote chain_id to BIGINT, add available_to_member_ids, and fix distribution status logic
Promotes chain_id from INT to BIGINT across schema, domain, and infra to support
millisecond-scale sender_key_version values generated by Rust. Adds available_to_member_ids
to the distribution status response so callers can distinguish already-available receivers
from those still needing a fresh upload. Fixes pending receivers to only include failed
distributions; available distributions now populate the new field instead.

+++ internal/application/e2ee/usecase/create_sender_key_request_test.go
@@ -171,7 +171,7 @@ func (s *senderKeyReqDistributionRepoStub) UpsertBatch(context.Context, []*sende
 	return nil
 }
 
-func (s *senderKeyReqDistributionRepoStub) FindPendingReceivers(context.Context, chatmember.ID, int) ([]chatmember.ID, error) {
+func (s *senderKeyReqDistributionRepoStub) FindPendingReceivers(context.Context, chatmember.ID, int64) ([]chatmember.ID, error) {
 	return nil, nil
 }
