+++ internal/domain/senderkeydistribution/repository.go
@@ -8,13 +8,13 @@ import (
 )
 
 type Repository interface {
-	// UpsertBatch records that receiverMemberID has fetched each sender's key at the given chain_id.
-	// Uses ON CONFLICT (sender_member_id, receiver_member_id) DO UPDATE to update chain_id if newer.
+	// UpsertBatch records that receiverMemberID has fetched each sender's key at the given sender-key version.
+	// The legacy chain_id mirror must stay aligned with sender_key_version.
 	UpsertBatch(ctx context.Context, dists []*SenderKeyDistribution) error
 
 	// FindPendingReceivers returns member IDs in the room who have not yet fetched
 	// senderMemberID's key at latestChainID (or have never fetched it at all).
-	FindPendingReceivers(ctx context.Context, senderMemberID chatmember.ID, latestChainID int) ([]chatmember.ID, error)
+	FindPendingReceivers(ctx context.Context, senderMemberID chatmember.ID, latestChainID int64) ([]chatmember.ID, error)
 
 	UpsertAvailable(ctx context.Context, dist *SenderKeyDistribution) error
 	FindLatest(ctx context.Context, senderMemberID, receiverMemberID chatmember.ID) (*SenderKeyDistribution, error)
