+++ internal/domain/senderkeydistribution/senderkeydistribution.go
@@ -6,8 +6,8 @@ import (
 	"github.com/HiroLiang/tentserv-chat-server/internal/domain/chatmember"
 )
 
-// SenderKeyDistribution records that receiver_member_id has fetched
-// sender_member_id's sender key at chain_id.
+// SenderKeyDistribution records the latest sender key distribution state for a sender/receiver pair.
+// ChainID is a legacy mirror of SenderKeyVersion and must track the same 64-bit version number.
 type SenderKeyDistribution struct {
 	ID                  ID
 	SenderMemberID      chatmember.ID
@@ -16,7 +16,7 @@ type SenderKeyDistribution struct {
 	SenderKeyVersion    int64
 	DistributionMessage []byte
 	Status              Status
-	ChainID             int
+	ChainID             int64
 	DistributedAt       time.Time
 	ConsumedAt          *time.Time
 	FailedAt            *time.Time
